added the function findMin in the code
and the program is complete here which is to find the minimum in the rotated array and the problem statement is:-  Given an integer array arr of size N, sorted in ascending order (with distinct values). Now the array is rotated between 1 to N times which is unknown. Find the minimum element in the array.